Moved several test cases to AppVeyor
We'll see if they still work...
